auto_trader/requirements.txt
# HTTP / API
requests>=2.31.0
aiohttp>=3.9.0
websockets>=12.0

# Scheduling
APScheduler>=3.10.4

# Configuration
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Data / Analysis
pandas>=2.1.0
numpy>=1.26.0

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# AI / LLM
openai>=1.12.0

# Discord Bot
discord.py>=2.3.0

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0

# Database
aiosqlite>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Utilities
pytz>=2024.1
